Partial revamp of communications
Replaced most communications to the phone with a new scheme: Pebble will only receive the initial message at the start, or if the settings on the phone are changed, and will not work until said message is received in full. Afterwards, the run will work INDEPENDENTLY on the pebble (while the option for a run on the phone is an option). When the run is done, the pebble will send the run information to the phone.