Fixed button press locks
Button presses will now do nothing if the pebble has not yet received the initial data from the phone or if it is currently sending run data to the phone, displaying a message to inform the user. It also resets the run after the run send has been completed, so that another run can be sent afterwards.